Tests/LibWeb/Text/expected/canvas/path2d-addpath-transform.txt

translated point: true
original point: false
